requirements.txt
google-adk==1.6.1
litellm==1.74.3
python-dotenv==1.0.0
